refactor: move functions and improve naming
