+++ tensilelite/Tensile/Common/__init__.py
@@ -2,4 +2,3 @@ from .Constants import *
 from .Parallel import *
 from .Types import *
 from .Utilities import *
-from .ArchVariant import *
